Manage Access catalog items
This is mainly on finding catalog items, even the ones hidden and delete them by uuid. There is an issue with mainly app sources where you create a new app source but rename it, that is not supported and the app source will exist but be hidden from the admin UI. Using the catalog search for SAML items with a name filter you can get the required uuid attribute and delete the hidden app source which will mess with access policies.